+++ src/pages/auth/AuthPasswordScreen.js
@@ -1,12 +1,12 @@
 import React, { useContext, useState } from "react";
 import { StyleSheet, View, TextInput } from "react-native";
-import { MainButton } from "../../MainButton";
 import { NavbarTop } from "../../components/NavbarTop";
 import { observer } from "mobx-react-lite";
 import { NotesStoreContext } from "../../store/NotesStore";
 import { Container } from "../../components/grid/Container";
 import { MainContainer } from "../../components/grid/MainContainer";
 import { useMessage } from "../../hooks/message.hook";
+import { MainButton } from "../../components/MainButton";
 
 export const AuthPasswordScreen = observer(({ navigation }) => {
   const [password, onPassword] = useState("123456");
